Remove some unneeded crymath files.
