+++ Code/Legacy/CryCommon/Cry_Geo.h
@@ -15,98 +15,9 @@
 // Forward declarations                                                      //
 ///////////////////////////////////////////////////////////////////////////////
 
-struct Ray;
-template <typename F>
-struct Lineseg_tpl;
-template <typename F>
-struct Triangle_tpl;
-
 struct AABB;
 template <typename F>
 struct OBB_tpl;
-
-//-----------------------------------------------------
-
-
-///////////////////////////////////////////////////////////////////////////////
-// Definitions                                                               //
-///////////////////////////////////////////////////////////////////////////////
-
-struct PosNorm
-{
-    Vec3    vPos;
-    Vec3    vNorm;
-};
-
-///////////////////////////////////////////////////////////////////////////////
-///////////////////////////////////////////////////////////////////////////////
-///////////////////////////////////////////////////////////////////////////////
-// struct Ray
-///////////////////////////////////////////////////////////////////////////////
-///////////////////////////////////////////////////////////////////////////////
-///////////////////////////////////////////////////////////////////////////////
-struct Ray
-{
-    Vec3 origin;
-    Vec3 direction;
-
-    //default Ray constructor (without initialisation)
-    inline Ray(void) {}
-    inline Ray(const Vec3& o, const Vec3& d) {  origin = o; direction = d; }
-    inline void operator () (const Vec3& o, const Vec3& d) {  origin = o; direction = d; }
-
-    ~Ray(void) {};
-};
-
-
-///////////////////////////////////////////////////////////////////////////////
-///////////////////////////////////////////////////////////////////////////////
-///////////////////////////////////////////////////////////////////////////////
-// struct Lineseg
-///////////////////////////////////////////////////////////////////////////////
-///////////////////////////////////////////////////////////////////////////////
-///////////////////////////////////////////////////////////////////////////////
-struct Lineseg
-{
-    Vec3_tpl<float> start;
-    Vec3_tpl<float> end;
-
-    //default Lineseg constructor (without initialisation)
-    inline Lineseg(void) {}
-    inline Lineseg(const Vec3_tpl<float>& s, const Vec3_tpl<float>& e) {  start = s; end = e; }
-
-    ~Lineseg(void) {};
-};
-
-///////////////////////////////////////////////////////////////////////////////
-///////////////////////////////////////////////////////////////////////////////
-///////////////////////////////////////////////////////////////////////////////
-// struct Triangle
-///////////////////////////////////////////////////////////////////////////////
-///////////////////////////////////////////////////////////////////////////////
-///////////////////////////////////////////////////////////////////////////////
-template <typename F>
-struct Triangle_tpl
-{
-    Vec3_tpl<F> v0, v1, v2;
-
-    //default Lineseg constructor (without initialisation)
-    ILINE Triangle_tpl(void) {}
-    ILINE Triangle_tpl(const Vec3_tpl<F>& a, const Vec3_tpl<F>& b, const Vec3_tpl<F>& c) {  v0 = a; v1 = b; v2 = c; }
-    ILINE void operator () (const Vec3_tpl<F>& a, const Vec3_tpl<F>& b, const Vec3_tpl<F>& c) { v0 = a; v1 = b; v2 = c; }
-
-    ~Triangle_tpl(void) {};
-
-    Vec3_tpl<F> GetNormal() const
-    {
-        return ((v1 - v0) ^ (v2 - v0)).GetNormalized();
-    }
-
-    F GetArea() const
-    {
-        return 0.5f * (v1 - v0).Cross(v2 - v0).GetLength();
-    }
-};
 
 ///////////////////////////////////////////////////////////////////////////////
 ///////////////////////////////////////////////////////////////////////////////
@@ -358,25 +269,3 @@ struct OBB_tpl
 
 typedef OBB_tpl<f32>    OBB;
 
-
-
-///////////////////////////////////////////////////////////////////////////////
-///////////////////////////////////////////////////////////////////////////////
-///////////////////////////////////////////////////////////////////////////////
-// struct Sphere
-///////////////////////////////////////////////////////////////////////////////
-///////////////////////////////////////////////////////////////////////////////
-///////////////////////////////////////////////////////////////////////////////
-struct Sphere
-{
-    Vec3 center;
-    float radius;
-
-    Sphere() {}
-    Sphere(const Vec3& c, float r)
-        : center(c)
-        , radius(r) {}
-    void operator()(const Vec3& c, float r) { center = c; radius = r; }
-};
-
-typedef Triangle_tpl<f32>   Triangle;
